feat: Adicionar animações de tela
