--- include/about.h
@@ -1,7 +1,0 @@
-#ifndef ABOUT_H
-#define ABOUT_H
-
-// Declaração das funções
-void about();
-
-#endif
