--- include/credits.h
@@ -1,7 +1,0 @@
-#ifndef CREDITS_H
-#define CREDITS_H
-
-// Declaração das funções
-void credits();
-
-#endif
