--- include/exit.h
@@ -1,7 +1,0 @@
-#ifndef EXIT_H
-#define EXIT_H
-
-// Declaração das funções
-int sair();
-
-#endif
