--- include/home.h
@@ -1,7 +1,0 @@
-#ifndef HOME_H
-#define HOME_H
-
-// Declaração das funções
-void home();
-
-#endif
