+++ main.c
@@ -1,10 +1,7 @@
 #include <stdio.h>
 #include <ncurses.h>
 #include <unistd.h>
-#include <home.h>
-#include <credits.h>
-#include <exit.h>
-#include <about.h>
+#include "screens.h"
 
 const char *PROJECT_VERSION = "1.0-alpha";
 
@@ -26,19 +23,8 @@ int main(void)
     // Fechar a janela com ctrl + c
     cbreak();
     
-    /* code */
-    // Caracteres para a animação
-    char* fade_chars[] = {"#", "+", ".", " "}; // do mais escuro ao mais claro
-
-    for(int f = 0; f < 4; f++) {
-        for(int y = 0; y < LINES; y++) {
-            for(int x = 0; x < COLS; x++) {
-                mvprintw(y, x, "%s", fade_chars[f]);
-            }
-        }
-        refresh();
-        usleep(500000);
-    }
+    // Animação de entrada
+    fade_animation("in");
 
     // Função de animação da borda
     draw_building_border('#',0, 0);
@@ -96,6 +82,9 @@ int main(void)
 
         resp = getch();
     }
+
+    // Animação de saida
+    fade_animation("out");
     
     // Finaliza ncurses
     endwin();
@@ -115,36 +104,5 @@ void draw_border(char style, int border_margin_left_right, int border_margin_top
     for (int y = border_margin_top_bottom; y < (LINES - border_margin_top_bottom); y++) {
         mvaddch(y, border_margin_left_right, style);               // borda esquerda
         mvaddch(y, COLS - 1 - border_margin_left_right, style);   // borda direita
-    }
-}
-
-void draw_building_border(char style, int border_margin_left_right, int border_margin_top_bottom)
-{
-    // Desenha topo gradualmente
-    for(int x = border_margin_left_right; x < COLS - border_margin_left_right; x++) {
-        mvaddch(border_margin_top_bottom, x, style);
-        refresh();
-        usleep(15000);
-    }
-
-    // Desenha direita gradualmente
-    for(int y = border_margin_top_bottom + 1; y < LINES - border_margin_top_bottom; y++) {
-        mvaddch(y, COLS - 1 - border_margin_left_right, style);
-        refresh();
-        usleep(15000);
-    }
-
-    // Desenha base gradualmente
-    for(int x = COLS - 2 - border_margin_left_right; x >= border_margin_left_right; x--) {
-        mvaddch(LINES - 1 - border_margin_top_bottom, x, style);
-        refresh();
-        usleep(15000);
-    }
-
-    // Desenha esquerda gradualmente
-    for(int y = LINES - 2 - border_margin_top_bottom; y > border_margin_top_bottom; y--) {
-        mvaddch(y, border_margin_left_right, style);
-        refresh();
-        usleep(15000);
     }
 }
